Move abstracted download func into utils
Before runs are downloaded, the indicator csv download code had to be
generalized and moved to utils.
