+++ api/src/utils.py
@@ -1,11 +1,14 @@
 import os
 import tempfile
 import time
+import csv
 from urllib.parse import urlparse
-from io import BytesIO
+from io import BytesIO, StringIO
+from zlib import compressobj
 
 from elasticsearch import Elasticsearch
 import boto3
+import pandas as pd
 
 from src.settings import settings
 from validation import ModelSchema
@@ -55,6 +58,46 @@ def delete_matching_records_from_model(model_id, record_key, record_test):
     modify_model(model_id, ModelSchema.ModelMetadataPatchSchema(**update))
 
     return record_count
+
+
+def download_csv_from_data_paths(data_paths, should_compress=True):
+
+    async def iter_csv():
+        # Build single dataframe
+        df = pd.concat(pd.read_parquet(file) for file in data_paths)
+
+        # Ensure pandas floats are used because vanilla python ones are problematic
+        df = df.fillna('').astype(
+            {col: 'str' for col in df.select_dtypes(include=['float32', 'float64']).columns},
+            # Note: This links it to the previous `df` so not a full copy
+            copy=False
+        )
+
+        # Prepare for writing CSV to a temporary buffer
+        buffer = StringIO()
+        writer = csv.writer(buffer)
+
+        # Write out the header row
+        writer.writerow(df.columns)
+
+        yield buffer.getvalue()
+        buffer.seek(0)  # To clear the buffer we need to seek back to the start and truncate
+        buffer.truncate()
+
+        # Iterate over dataframe tuples, writing each one out as a CSV line one at a time
+        for record in df.itertuples(index=False, name=None):
+            writer.writerow(str(i) for i in record)
+            yield buffer.getvalue()
+            buffer.seek(0)
+            buffer.truncate()
+
+    async def compress(content):
+        compressor = compressobj()
+        async for buff in content:
+            yield compressor.compress(buff.encode())
+        yield compressor.flush()
+
+    return compress(iter_csv()) if should_compress else iter_csv()
 
 
 def run_model_with_defaults(model_id):
